(Feature) Torus login #1441 (#1528)
* add: torus files

* fix: typo

* fix: react warning

* fix: react warning on unmounted state change

* add: basic torus support

* add: use torus private key as hdwallet seed

* add: torus new envs

* add: upgrade torus sdk

* add: register torus servie worker to servce redirect page

* add: redirect is served from service worker

* add: finalize torus. protect signup routes

* fix: make sure token exists before deleting w3 user

* refactor: torus auth

* refactor: indexeddb call

* fix: typo

* add: misc fixes and additions

* add: gundb optimizations and hacks

* add: handle existing user torus login

* add: refactor signupstate with torus

* add: (#1515) Revise torus login page to match required design

* add: facebook login

* fix: gun hacks so it doesnt get stuck

* fix: enough delay so gun can respond

+++ src/lib/hooks/useDeleteAccountDialog.js
@@ -44,8 +44,15 @@ export default ({ API, showDialog, store, theme }) =>
                 log.debug('deleted account', isDeleted)
 
                 if (isDeleted) {
-                  API.deleteWalletFromW3Site(token)
-                  await Promise.all([AsyncStorage.clear()])
+                  token && API.deleteWalletFromW3Site(token)
+                  const req = new Promise((res, rej) => {
+                    const del = indexedDB.deleteDatabase('radata')
+                    del.onsuccess = res
+                    del.onerror = rej
+                  })
+
+                  //remove all local data so its not cached and user will re-login
+                  await Promise.all([AsyncStorage.clear(), req.catch()])
                   window.location = '/'
                 } else {
                   showDialog('There was a problem deleting your account. Try again later.')
